# LostAndFound/src/main/resources/api.yml
openapi: 3.0.0
info:
  title: Lost and Found API
  version: 1.0.0
  description: API for lost and found items managing.

servers:
  - url: http://localhost:8080/api  # Replace with your server URL

paths:
  /admin/upload-items:
    post:
      summary: Upload lost items from a file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Text file containing lost item data (ItemName, Quantity, Place)
      responses:
        '200':
          description: Lost items uploaded successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LostItem'
        '500':
          description: Internal server error

  /lost-items:
    get:
      summary: Retrieve all lost items
      responses:
        '200':
          description: List of lost items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LostItem'

  /claim-item:
    post:
      summary: Claim a lost item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Claim'
      responses:
        '200':
          description: Claim created successfully
          content:
            application/json:
              schema:
               type: string
               example: "Claim report successfully."
        '400':
          description: Bad Request (e.g., The claim item is not existing)

  /admin/check-claims:
    get:
      summary: Retrieve all claims
      responses:
        '200':
          description: List of claims
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Claim'

  /admin/check-user-claims:
    get:
      summary: Retrieve all claims from a specific user
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
            format: int64
          required: true
          description: Numeric ID of the user
      responses:
        '200':
          description: List of claims
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Claim'


components:
  schemas:
    LostItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
        itemName:
          type: string
        quantity:
          type: integer
          format: int32
        place:
          type: string
    Claim:
      type: object
      required:
        - userId
        - itemId
        - claimedQuantity
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        itemId:
          type: integer
          format: int64
        claimedQuantity:
          type: integer
          format: int32
        userName:
          type: string